requirements.txt
# === Librerías necesarias para ejecutar el sistema ===
customtkinter==5.2.2
pillow==10.3.0
supabase-py==2.4.0


gotrue==2.4.2
postgrest==0.14.3
storage3==0.6.1
realtime==1.0.2


